+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## Unreleased
+
+- Renamed `ResourceGroup` parameter to `ResourceGroupName` in
+  `New-CosmosDbContext` function - fixes [Issue #158](https://github.com/PlagueHO/CosmosDB/issues/158).
+- Correct `*-CosmosDbAccount` functions examples in README.MD to show
+  `ResourceGroupName` parameter.
+- Added `Get-CosmosDbAccountMasterKey` function for retrieving the keys
+  of an existing account in Azure - fixes [Issue #162](https://github.com/PlagueHO/CosmosDB/issues/162).
+- Added `New-CosmosDbAccountMasterKey` function for regenerating the keys
+  of an existing account in Azure - fixes [Issue #164](https://github.com/PlagueHO/CosmosDB/issues/164).
+
 ## 2.1.10.103
 
 - Added support for creating and updating documents containing